Extracted QueryBuilder data source from abstract Mapped data source, created new abstract class DataSource, moved Counting AST Walker to Utils
